requirements.txt
streamlit==1.12.0
pandas==1.5.0
mysql-connector-python==8.0.31
fuzzywuzzy==0.18.0
SQLAlchemy==1.4.41
python-Levenshtein==0.12.2
